select random question from firebase but can't reload current page when choice answer
