+++ src/actions/quesAction.js
@@ -1,6 +1,7 @@
-export const createQues = (ques) =>{
+export const getQues = (ques) =>{
     return (dispatch, getState, { getFirebase, getFirestore }) => {
         //make async to database
+        
         dispatch({ 
             type : 'GET_QUES', 
             ques
