+++ src/component/PlayerLayout/PlayerLayout.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import { Container, Row} from 'reactstrap'
+import { Container, Row, Col} from 'reactstrap'
 import Web3 from 'web3'
 import { connect } from 'react-redux'
 import { firestoreConnect } from 'react-redux-firebase'
@@ -39,26 +39,45 @@ class PlayerLayout extends Component {
     render() {   
         const { rank } = this.props;
         const { question } = this.props;
-        // console.log(question);
-        return (   
-            <div>
-                <Container>
-                    <Row className="set_height">
-                        {question && question.map(ques => (
-                            // ques.id= '9VjqRPa9EzCARyXCB0EA'
-                             <QuesArea key = {ques.id} ques ={ques} acc={this.state}/> 
-                        ))}
-                        <RankArea rank = {rank} />
-                    </Row>
-                </Container>
-            </div>
-        );
+        if(question != null){
+            return (   
+                <div>
+                    <Container>
+                        <Row className="set_height">
+                                <QuesArea ques ={question} acc={this.state}/> 
+                            <RankArea rank = {rank} />
+                        </Row>
+                    </Container>
+                </div>
+            );
+        }else{
+            return (   
+                <div>
+                    <Container>
+                        <Row className="set_height">
+                            <Col className = "box_color" xs="8">
+                                <div className="margin_box ">
+                                    <h1> Waiting ...</h1>
+                                </div>
+                            </Col>
+                            <RankArea rank = {rank} />
+                        </Row>
+                    </Container>
+                </div>
+            );
+        }
+        
     }
 }
 
 const mapStatetoProps = (state) => {
+    const question = state.firestore.data.project_hunter
+    const id = Math.floor(Math.random()*10);
+    const quesID =state.ques.questionID[id];
+    const ques = question ? question[quesID] : null
+    
     return {
-        question : state.firestore.ordered.project_hunter,
+        question : ques,
         rank : state.rank.ranking
     }
 }
@@ -67,8 +86,7 @@ export default compose(
     connect(mapStatetoProps),
     firestoreConnect([
         { 
-            collection : 'project_hunter',
-            doc : '9VjqRPa9EzCARyXCB0EA'      
+            collection : 'project_hunter',     
         }
     ])
 )(PlayerLayout);
